Disgusting hack to save team/enemy skins
InitClientPersistent() is being called on every respawn. This function
sets the entire client_persistent_t struct back to zero (line 1110).
This struct is supposed to survive respawns and even gamemaps. This
should all be sorted out eventually, but for now just save the skins and
restore them after the struct is cleared.